+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 2.1.0 (2022-03-12)
 
 - Feat: added `unplugin-auto-import` for automatic package imports.
+- Feat: added `cssnano` for minimizing production CSS.
 - Chore: bumped deps.
 
 ## 2.0.0 (2022-03-06)